+++ test/tempo/interval_test.exs
@@ -1,0 +1,6 @@
+defmodule Tempo.IntervalTest do
+  use ExUnit.Case, async: true
+  import Tempo.Sigils
+
+  doctest Tempo.Interval
+end